command[1]/command/requirements.txt
africastalking==1.2.4
aiohttp==3.8.1
aiohttp-jinja2==1.5
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.2.0
blinker==1.4
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
contextlib2==21.6.0
distlib==0.3.4
Django==3.2.12
dnspython==2.2.1
dominate==2.6.0
email-validator==1.1.3
filelock==3.6.0
Flask==2.0.2
Flask-Bootstrap4==4.0.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-Toastr==0.5.7
Flask-WTF==1.0.1
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
platformdirs==2.5.1
pycryptodome==3.15.0
PyMySQL==1.0.2
pytz==2022.1
requests==2.27.1
schedule==1.1.0
schema==0.7.5
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
tinydb==4.7.0
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.14.0
virtualenvwrapper-win==1.2.7
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.1
yarl==1.7.2
zipp==3.7.0
